Revert "What about software freedom?"
This reverts commit a419f9b2cfdcfab63fd5ef5ba080512a50b64b6b.

Per email discussion, this was probably just a test commit. In any case,
it breaks the line length on the default Chrome width (it spills to two
lines), so reverting it.

If someone cares about mentioning "free" here, it should probably be
done in a way that explains what it means, since nobody will understand
it otherwise.

(And probably this isn't the right page to explain it on, anyway.)